requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
psycopg2-binary==2.9.7
supabase==1.3.0
asyncpg==0.29.0

# Pydantic for data validation
pydantic==2.4.2
pydantic-settings==2.0.3

# Environment management
python-dotenv==1.0.0

# Image processing and AI
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3

# HTTP requests and API clients
requests==2.31.0
httpx==0.25.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
typing-extensions==4.8.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1